requirements.txt
# sentence-transformers
Django
djangorestframework
openai
faiss-cpu

PyPDF2
python-docx
markdown
python-decouple
redis
django-cors-headers


